separate jdbc-comparison tags due to junit5-plugin bug which doesn't filters method-level tags
